Feat: 초대코드 재설정을 메뉴의 초대코드란으로 통일
