+++ components/organisms/AgitManageForm.tsx
@@ -1,13 +1,11 @@
 "use client";
 
-import { reissueInviteCodeAction, updateAgitAction } from "@/actions/agitActions";
+import { updateAgitAction } from "@/actions/agitActions";
 import { SubmitButton } from "@/components/atoms";
 import { AuthField, CapacityStepper } from "@/components/molecules";
 import { ThumbnailUpload } from "@/components/molecules/ThumbnailUpload";
 import { toast } from "@/components/ui/toast";
 import { ROUTES } from "@/config/routes";
-import { copyText } from "@/lib/copyText";
-import { Check, Copy, Link2 } from "lucide-react";
 import {
   AGIT_DEFAULT_MAX_CAPACITY,
   AGIT_DESCRIPTION_MAX_LENGTH,
@@ -29,9 +27,6 @@ export function AgitManageForm({ agit }: AgitManageFormProps) {
   );
   const [pending, setPending] = useState(false);
   const [error, setError] = useState<string | null>(null);
-  const [inviteCode, setInviteCode] = useState(agit.inviteCode?.trim() ?? "");
-  const [copied, setCopied] = useState(false);
-  const [reissuing, setReissuing] = useState(false);
 
   async function handleSubmit(formData: FormData) {
     if (pending) return;
@@ -60,30 +55,6 @@ export function AgitManageForm({ agit }: AgitManageFormProps) {
     router.refresh();
   }
 
-  async function copyInviteCode() {
-    if (!inviteCode) return;
-    const ok = await copyText(inviteCode);
-    setCopied(ok);
-    if (ok) {
-      toast.add({ type: "success", title: "초대코드를 복사했습니다" });
-    }
-  }
-
-  async function reissueInviteCode() {
-    if (reissuing) return;
-    setReissuing(true);
-    const result = await reissueInviteCodeAction(agit.id);
-    setReissuing(false);
-    if (!result.ok) {
-      toast.add({ type: "error", title: "초대코드를 재설정하지 못했습니다", description: result.error });
-      return;
-    }
-    setInviteCode(result.data);
-    setCopied(false);
-    toast.add({ type: "success", title: "초대코드를 재설정했습니다" });
-    router.refresh();
-  }
-
   return (
     <form className="flex w-full flex-col gap-3.5" action={handleSubmit}>
       <ThumbnailUpload />
@@ -107,31 +78,6 @@ export function AgitManageForm({ agit }: AgitManageFormProps) {
         defaultValue={agit.description}
         maxLength={AGIT_DESCRIPTION_MAX_LENGTH}
       />
-
-      <p className="m-0 text-[14px] font-medium text-[var(--dl-color-text-primary)]">초대코드</p>
-      <button
-        type="button"
-        className="flex min-h-[32px] w-full cursor-pointer items-center justify-between gap-2 rounded-[10px] border border-[#e3e0ed] bg-[#fff] p-[6px_10px] text-left disabled:cursor-default"
-        onClick={copyInviteCode}
-        disabled={!inviteCode}
-        aria-label="초대코드 복사"
-      >
-        <div className="flex min-w-0 items-center gap-2">
-          <Link2 className="size-3.5 shrink-0 text-[#262433]" strokeWidth={2} />
-          <p className="m-0 overflow-hidden text-xs font-medium tracking-[0.04em] text-[#262433] whitespace-nowrap [text-overflow:ellipsis]">
-            {inviteCode || "초대코드"}
-          </p>
-        </div>
-        {copied ? (
-          <Check className="size-3.5 shrink-0 text-[var(--dl-color-text-brand)]" strokeWidth={2} aria-hidden />
-        ) : (
-          <Copy className="size-3.5 shrink-0 text-[var(--dl-color-text-brand)]" strokeWidth={2} aria-hidden />
-        )}
-        <span className="sr-only">{copied ? "복사됨" : "복사"}</span>
-      </button>
-      <SubmitButton type="button" variant="outline" disabled={reissuing} onClick={reissueInviteCode}>
-        {reissuing ? "재설정 중..." : "초대코드 재설정"}
-      </SubmitButton>
 
       <p className="m-0 text-[14px] font-medium text-[var(--dl-color-text-primary)]">최대 인원</p>
       <div className="flex min-h-[68px] items-center justify-between gap-[12px] rounded-[12px] bg-[var(--dl-color-bg-brand-subtle)] p-[13px_14px]">
